% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R, R/omics.R
\name{install_easyMRomics}
\alias{install_easyMRomics}
\title{install easyMRomics package}
\usage{
install_easyMRomics()

install_easyMRomics()
}
\description{
install easyMRomics package

install easyMRomics package
}
\examples{
install_easyMRomics()
install_easyMRomics()
}
