% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package.R
\name{install_package_MR}
\alias{install_package_MR}
\title{install package for easyMR}
\usage{
install_package_MR(package)
}
\arguments{
\item{package}{}
}
\description{
install package for easyMR
}
\examples{
install_package_MR('gwasvcf')
}
